Task 10. Safe divide

+++ MySQL_Advanced/10-div.sql
@@ -9,7 +9,7 @@ BEGIN
     IF b = 0 THEN
         RETURN 0;
     END IF;
-    RETURN a / b;
+    RETURN FLOAT(a) / FLOAT(b);
 END;
 //
